unrelated: move asia module and split in submodules
