+++ workalendar/asia/__init__.py
@@ -1,0 +1,17 @@
+# -*- coding: utf-8 -*-
+from .japan import Japan
+from .malaysia import Malaysia
+from .qatar import Qatar
+from .singapore import Singapore
+from .south_korea import SouthKorea
+from .taiwan import Taiwan
+
+
+__all__ = (
+    Japan,
+    Malaysia,
+    Qatar,
+    Singapore,
+    SouthKorea,
+    Taiwan,
+)
